refactor: file-helper to file-context
