+++ jest.config.js
@@ -2,7 +2,7 @@ module.exports = {
   projects: [
     '<rootDir>/libs/adapter',
     '<rootDir>/libs/domain',
-    '<rootDir>/libs/file-helper',
+    '<rootDir>/libs/file-context',
     '<rootDir>/libs/vue',
   ],
 };
