+++ libs/adapter/index.ts
@@ -5,7 +5,6 @@ export * from './in-memory/in-memory.entity-repository';
 export * from './in-memory/in-memory.feature-service';
 
 export * from './mongo-db/mongo-db.entity-repository';
-export * from './mongo-db/mongo-db.file-store';
 export * from './mongo-db/mongo-db.helper';
 export * from './mongo-db/mongo-db.logging-service';
