+++ libs/file-context/index.ts
@@ -2,3 +2,4 @@ export * from './lib/file.context';
 export * from './lib/file.entity';
 export * from './lib/file-store';
 export * from './lib/id-should-be-unique.exception';
+export * from './lib/mongo-db.file-store';
