+++ libs/file-context/lib/mongo-db.file-store.ts
@@ -1,5 +1,7 @@
 import { Collection } from 'mongodb';
-import { FileEntity, FileStore, IdShouldBeUniqueException } from '@pl-oss/file-helper';
+import { FileStore } from './file-store';
+import { FileEntity } from './file.entity';
+import { IdShouldBeUniqueException } from './id-should-be-unique.exception';
 
 export class MongoDBFileStore implements FileStore {
   constructor(private readonly collection: Collection) {}
@@ -9,7 +11,6 @@ export class MongoDBFileStore implements FileStore {
       const withMongoId = { ...fileEntity, _id: fileEntity.id };
       await this.collection.insertOne(withMongoId);
     } catch (e) {
-      console.log(e);
       const DUPLICATE_KEY_ERROR_CODE = 11000;
       if (e?.code === DUPLICATE_KEY_ERROR_CODE) throw new IdShouldBeUniqueException();
     }
