+++ libs/file-context/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pl-oss/file-helper",
+  "name": "@pl-oss/file-context",
   "version": "0.0.52",
   "license": "MIT"
 }
