+++ libs/file-context/tsconfig.lib.json
@@ -1,9 +1,13 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
+    "module": "commonjs",
     "outDir": "../../dist/out-tsc",
     "declaration": true,
-    "types": []
+    "types": [],
+    "lib": [
+      "es2019"
+    ]
   },
   "include": [
     "**/*.ts"
