+++ libs/file-context/tsconfig.spec.json
@@ -3,13 +3,14 @@
   "compilerOptions": {
     "outDir": "../../dist/out-tsc",
     "module": "commonjs",
-    "types": ["jest", "node"]
+    "types": [
+      "jest",
+      "node"
+    ]
   },
   "include": [
     "**/*.spec.ts",
-    "**/*.spec.tsx",
     "**/*.spec.js",
-    "**/*.spec.jsx",
     "**/*.d.ts"
   ]
 }
