--- libs/file-helper/README.md
@@ -1,1 +1,0 @@
-# File Helper
