+++ nx.json
@@ -30,7 +30,7 @@
     "domain": {
       "tags": []
     },
-    "file-helper": {
+    "file-context": {
       "tags": []
     },
     "vue": {
