+++ tsconfig.base.json
@@ -24,8 +24,8 @@
       "@pl-oss/domain": [
         "libs/domain/index.ts"
       ],
-      "@pl-oss/file-helper": [
-        "libs/file-helper/src/index.ts"
+      "@pl-oss/file-context": [
+        "libs/file-context/index.ts"
       ],
       "@pl-oss/vue": [
         "libs/vue/index.ts"
